Split some code out of karabo_data.reader
